// src/app/enums/menu-main.enum.ts
export enum MenuMainEnum {
    ADULTS = 'ADULTS',
    CHANNEL = 'CHANNEL',
    CONTINUE_WATCHING = 'CONTINUE_WATCHING',
    DOCUMENTARIES = 'DOCUMENTARIES',
    FAVORITES = 'FAVORITES',
    HOME = 'HOME',
    MOVIES = 'MOVIES',
    RECHARGE = 'RECHARGE',
    SEARCH = 'SEARCH',
    SETTINGS = 'SETTINGS',
    SERIES = 'SERIES',
    SHOWS = 'SHOWS',
    STAND_UP = 'STAND_UP'
}